slew rate stuff
I double checked the slew rate code and triple checked it with chat gpt, and it looks like it should all be working correctly.

We really need to diagnose why the chassis was having weird driving issues. First we should get an ACTUALLY GOOD BATTERY and double check the joystick inputs on the dashboard. If it's still having driving issues then idk.